+++ apps/users/urls.py
@@ -1,0 +1,8 @@
+from django.contrib import admin
+from django.urls import path, include
+from . import views
+
+urlpatterns = [
+    path('signup', views.signup),
+    path('login', views.login)
+]
