+++ server/urls.py
@@ -1,10 +1,11 @@
 from django.contrib import admin
-from django.urls import re_path, path
+from django.urls import re_path, path, include
 from . import views
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    re_path('login', views.login),
-    re_path('signup', views.signup),
-    re_path('test', views.test)
+    path('account/', include('apps.users.urls')),
+    # re_path('login', views.login),
+    # re_path('signup', views.signup),
+    # re_path('test', views.test)
 ]
