+++ server/views.py
@@ -1,48 +1,48 @@
-from rest_framework.decorators import api_view
-from rest_framework.response import Response
-from django.http import JsonResponse
+# from rest_framework.decorators import api_view
+# from rest_framework.response import Response
+# from django.http import JsonResponse
 
-from django.shortcuts import get_object_or_404
-from .models import UserSerializer
-from rest_framework import status
-from rest_framework.authtoken.models import Token
-# from django.contrib.auth.models import User
-from .models import User
-from django.contrib.auth import authenticate
+# from django.shortcuts import get_object_or_404
+# from .models import UserSerializer
+# from rest_framework import status
+# from rest_framework.authtoken.models import Token
+# # from django.contrib.auth.models import User
+# from .models import User
+# from django.contrib.auth import authenticate
 
-from django.views.decorators.csrf import ensure_csrf_cookie
+# from django.views.decorators.csrf import ensure_csrf_cookie
 
-from rest_framework.decorators import api_view, authentication_classes, permission_classes
-from rest_framework.authentication import SessionAuthentication, TokenAuthentication
-from rest_framework.permissions import IsAuthenticated
+# from rest_framework.decorators import api_view, authentication_classes, permission_classes
+# from rest_framework.authentication import SessionAuthentication, TokenAuthentication
+# from rest_framework.permissions import IsAuthenticated
 
-@api_view(['POST'])
-def login(request):
-    # print(request.data)
-    user = get_object_or_404(User, username=request.data['username'].lower())
-    if not user.check_password(request.data['password']):
-        return JsonResponse({'loginExists': False})
-        # return Response("missing user", status=status.HTTP_404_NOT_FOUND)
-    token, created = Token.objects.get_or_create(user=user)
-    serializer = UserSerializer(user)
-    return JsonResponse({'loginExists': True})
-    # return Response({'token': token.key, 'user': serializer.data})
+# @api_view(['POST'])
+# def login(request):
+#     # print(request.data)
+#     user = get_object_or_404(User, username=request.data['username'])
+#     if not user.check_password(request.data['password']):
+#         return JsonResponse({'loginExists': False})
+#         # return Response("missing user", status=status.HTTP_404_NOT_FOUND)
+#     token, created = Token.objects.get_or_create(user=user)
+#     serializer = UserSerializer(user)
+#     return JsonResponse({'loginExists': True})
+#     # return Response({'token': token.key, 'user': serializer.data})
 
-@api_view(['POST'])
-def signup(request):
-    serializer = UserSerializer(data=request.data)
-    # print(request.data)
-    if serializer.is_valid():
-        serializer.save()
-        user = User.objects.get(username=request.data['username'].lower())
-        user.set_password(request.data['password'])
-        user.save()
-        token = Token.objects.create(user=user)
-        return Response({'token': token.key, 'user': serializer.data})
-    return Response(serializer.errors, status=status.HTTP_200_OK) 
+# @api_view(['POST'])
+# def signup(request):
+#     serializer = UserSerializer(data=request.data)
+#     # print(request.data)
+#     if serializer.is_valid():
+#         serializer.save()
+#         user = User.objects.get(username=request.data['username'])
+#         user.set_password(request.data['password'])
+#         user.save()
+#         token = Token.objects.create(user=user)
+#         return Response({'token': token.key, 'user': serializer.data})
+#     return Response(serializer.errors, status=status.HTTP_200_OK) 
 
-@api_view(['GET'])
-@authentication_classes([SessionAuthentication, TokenAuthentication])
-@permission_classes([IsAuthenticated])
-def test(request):
-    return Response('Passed for {}'.format(request.user.email))
+# @api_view(['GET'])
+# @authentication_classes([SessionAuthentication, TokenAuthentication])
+# @permission_classes([IsAuthenticated])
+# def test(request):
+#     return Response('Passed for {}'.format(request.user.email))
